valid address generics when add minter nft template
